Implemented startup, peer and orderer creation.
