+++ blockchain/controllers/test.js
@@ -1,4 +1,4 @@
-import blockchainController from './blockChainController.js';
+import blockchainController from './blockChainFunctionHandler.js';
 
 const controller = new blockchainController();
 // controller.startChain();
@@ -11,4 +11,7 @@ const controller = new blockchainController();
 // controller.install_chaincode_metadata('asset_1.0','../generated_resources/chaincode.tgz', ['org1','org2'], '../generated_resources/network-config.yaml');
 // controller.deploy_external_chaincode('asset','vparash0299/asset-transfer-basic-ts','../generated_resources/network-config.yaml');
 // controller.approve_chaincode('asset','1.0',1,'demo',['org1','org2'],'../generated_resources/network-config.yaml');
-controller.commit_chaincode('asset','1.0',1,'demo',['org1','org2'],'../generated_resources/network-config.yaml');
+// controller.commit_chaincode('asset','1.0',1,'demo',['org1','org2'],'../generated_resources/network-config.yaml');
+
+import { createBlockchain } from './blockChainController.js';
+createBlockchain();
