+++ router/blockchainRouter.js
@@ -1,0 +1,7 @@
+import express from 'express';
+import { createBlockchain } from '../blockchain/controllers/blockChainController.js';
+const blockchainRouter = express.Router();
+
+blockchainRouter.post('/start',  createBlockchain);   
+
+export default blockchainRouter;
